Add PXPaymentFlowModel
